multi_usrp: unify GPIO access type
GPIOs in the property tree are registered as uint32_t's, so the
get_gpio_attr function should use that type as well.

This resolves a property tree runtime_error when running the `gpio`
example with a B2xx device.

+++ host/lib/usrp/multi_usrp.cpp
@@ -2218,7 +2218,7 @@ public:
                         return val;
                     }
                     default:
-                        return uint32_t(_tree->access<uint64_t>(
+                        return uint32_t(_tree->access<uint32_t>(
                             mb_root(mboard) / "gpio" / bank / attr).get());
                 }
                 return 0;
